backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.12.0
cachetools==5.3.2
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.0
cssselect2==0.7.0
defusedxml==0.7.1
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
Faker==23.1.0
filelock==3.13.1
flake8==7.0.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-Session==0.6.0
fonttools==4.49.0
fpdf2 @ git+https://github.com/Clock-Speed/fpdf2.git@4b0d537cc24ca51858206e3824b46be8f63c597b
frozenlist==1.4.1
gptcache==0.1.43
guidance==0.1.10
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
lxml==5.1.0
MarkupSafe==2.1.4
mccabe==0.7.0
mongomock==4.1.2
msal==1.26.0
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.9.0
opencv-python==4.9.0.80
ordered-set==4.1.0
packaging==23.2
paramiko==3.4.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
pyflakes==3.2.0
pyformlang==1.0.4
PyJWT==2.8.0
pymongo==4.6.2
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.1
pypdf==4.0.1
pypng==0.20220715.0
pyproject-api==1.6.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
qrcode==7.4.2
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
sentinels==1.0.0
six==1.16.0
sniffio==1.3.0
sshtunnel==0.1.4
svglib==1.5.1
svgwrite==1.4.3
tiktoken==0.5.2
tinycss2==1.2.1
tox==4.12.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
webencodings==0.5.1
Werkzeug==3.0.1
wonderwords==2.2.0
yarl==1.9.4
zipp==3.17.0
